requirements.txt
sklearn
matplotlib
seaborn
numpy
pandas
optuna
neptune-client[optuna]
torch
torchvision